Update demo tests to check for success
Currently the tests which intend to check demos build correctly
for different configuration files/formats assert the first
error thrown is not a given error. This commit updates these
tests to confirm no errors are thrown at all by providing missing
mustache templates and the demos we expect to be built are.

All fixtures are reset after each test to help ensure one test
does not interfere with another.